+++ src/index.js
@@ -11,5 +11,5 @@ export {
   extractPrefix, classifyPrefix,
   NODE_ID_REGEX, NODE_ID_MAX_LENGTH, CANONICAL_PREFIXES, SYSTEM_PREFIXES, ALL_PREFIXES,
 } from './validators.js';
-export { defineView, renderView, listViews } from './views.js';
+export { defineView, declareView, renderView, listViews } from './views.js';
 export { parseDirectives, processCommit } from './hooks.js';